Fix issue with Ember Inflector deprecation warning
When we use the public ember-inflector v 3.0.1 we get an issue
about modifying the Ember.Inflector object.

Not sure why this happens, but simply dropping this code makes
it go away and since this code will be removed at some point anyway
we can run off of this fork until then.